requirements.txt
ragas==0.1.8
llama-index-readers-web
python-dotenv
deepeval
